App.tsx is mow 16 lines
