+++ src/components/feed/index.ts
@@ -1,0 +1,2 @@
+export * from './PostCard';
+export * from './Timeline';
